backend/requirements.txt
# backend/requirements.txt - For Production

Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-JWT-Extended
Flask-Bcrypt
Flask-Mail
Flask-Limiter
Werkzeug
SQLAlchemy
psycopg2-binary
python-dotenv
marshmallow
marshmallow-sqlalchemy
Pillow
requests
celery
redis
gunicorn
python-dateutil
pandas
numpy
scikit-learn
beautifulsoup4
selenium
webdriver-manager